Revert "add Keybinds"
