+++ Cargo.toml
@@ -6,6 +6,7 @@ description = "A modern, animated dock for Hyprland built with Rust and Bevy Eng
 readme = "README.md"
 repository = "https://github.com/horberlan/anny-dock"
 license = "MIT"
+license-file = "LICENSE"
 keywords = ["dock", "bevy", "hyprland", "animated"]
 authors = ["pois.nada"]
 
@@ -16,10 +17,10 @@ resources = ["assets/"]
 paths = ["assets/"]
 
 [dependencies]
-bevy = "0.12"
-bevy_easings = "0.12"
-bevy_embedded_assets = "0.9"
-bevy_svg = { version = "0.12", default-features = false, features = ["2d"] }
+bevy = "0.11"
+bevy_easings = "0.11"
+bevy_embedded_assets = "0.8"
+bevy_svg = { version = "0.11", default-features = false, features = ["2d"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 image = "0.24"
