+++ src/systems/icon.rs
@@ -73,6 +73,7 @@ pub fn reorder_icons_system(
 
     reorder_trigger.0 = false;
 }
+
 fn lerp(start: f32, end: f32, t: f32) -> f32 {
     start + (end - start) * t
 }
