+++ src/systems/mod.rs
@@ -1,16 +1,14 @@
-pub mod animation;
-mod camera;
 mod drag;
 mod icon;
-mod scroll;
-mod keybinds;
 mod title;
+mod camera;
+mod scroll;
+pub mod animation;
 
-pub use animation::icon_scale_animation_system;
-pub use animation::ScrollAnimationState;
-pub use camera::*;
 pub use drag::*;
 pub use icon::*;
-pub use scroll::*;
-pub use keybinds::*;
 pub use title::*;
+pub use camera::*;
+pub use scroll::*;
+pub use animation::ScrollAnimationState;
+pub use animation::icon_scale_animation_system;
